\name{S1.gpr}
\alias{S1.gpr, S2.gpr, S4.gpr}
\docType{data}
\title{ GenePix result files}
\description{

    GenePix result files are tab delimited text files exported from the commonly
    used microarray image analysis tool GenePix.

}

\format{
  tab delimeted text file
}

\source{
 The GenePix result files are files from original reverse phase protein arrays
}


\keyword{datasets}
